% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_objects.R
\docType{data}
\name{or_mask}
\alias{or_mask}
\title{A mask for Oregon's exterior}
\format{An object of type \link[sf:sf]{sf::sf}}
\usage{
or_mask
}
\description{
A spatial object containing a mask of Oregon's near exterior. This was derived from
the \href{https://spatialdata.oregonexplorer.info/geoportal/details;id=cf375eb0ff7042d99daea17a776821a8}{2001 Oregon State Boundary shapefile}.
}
\keyword{datasets}
