% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_objects.R
\docType{data}
\name{or_signed_routes}
\alias{or_signed_routes}
\title{Signed routes in Oregon}
\format{An object of type \link[sf:sf]{sf::sf}}
\source{
\url{ftp://ftp.odot.state.or.us/tdb/trandata/GIS_data/Transportation_Network/signed_rtes.zip}
}
\usage{
or_signed_routes
}
\description{
A spatial object containing lines for each signed route in Oregon.
}
\keyword{datasets}
