% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_objects.R
\docType{data}
\name{or_water_bodies}
\alias{or_water_bodies}
\title{Water bodies in Oregon}
\format{An object of type \link[sf:sf]{sf::sf}}
\source{
\url{https://spatialdata.oregonexplorer.info/geoportal/details;id=3439a3c43f9f4c4499802f55898b7dd8}
}
\usage{
or_water_bodies
}
\description{
A spatial object containing area data for large water bodies in Oregon.
}
\keyword{datasets}
