% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oregon_mask.R
\name{oregon_mask}
\alias{oregon_mask}
\title{Masks out the area just outside Oregon with a solid color.}
\usage{
oregon_mask(road_size = 0.01, road_color = NA, water_color = NA,
  bg_color = "#222222", xlim = c(-124.6, -116.4), ylim = c(41.9,
  46.3))
}
\value{
A list of ggplot2 geoms of types \code{\link[ggplot2:geom_sf]{ggplot2::geom_sf()}} and \code{\link[ggplot2:coord_sf]{ggplot2::coord_sf()}}.
}
\description{
Include this function in a ggplot2 sequence to plot a mask of the area outside Oregon.
Also includes the option to draw roads and bodies of water.
}
\details{
\itemize{
\item \code{road_size}: The thickness of the road lines.
\item \code{road_color}: The color of the road lines. When set to NA (the default),
roads are not rendered. Set this to any color to include roads in the plot.
\item \code{water_color}: The color of the water. When set to NA (the default), water
is not rendered. Set this to any color to include roads in the plot. Water
may take quite a while to render.
\item \code{bg_color}: The color of the mask surrounding Oregon.
\item \code{xlim}: Limits of the x-axis. Defaults to include all of Oregon.
\item \code{ylim}: Limits of the y-axis. Defaults to include all of Oregon.
}
}
\examples{
ggplot() + oregon_mask()

}
